extracted authentication to separate file
